chore: extract items into a separate file
