+++ items.ts
@@ -1,0 +1,188 @@
+export enum Item {
+  Nothing = 'Nothing',
+  Adena = 'Адена',
+
+  CloakEnchant = 'Свиток: Модифицировать Плащ',
+  Cloak = 'Плащ',
+  Cloak_1 = 'Плащ +1',
+  Cloak_2 = 'Плащ +2',
+  Cloak_3 = 'Плащ +3',
+  Cloak_4 = 'Плащ +4',
+  Cloak_5 = 'Плащ +5',
+  Cloak_6 = 'Плащ +6',
+  Cloak_7 = 'Плащ +7',
+  Cloak_8 = 'Плащ +8',
+  Cloak_9 = 'Плащ +9',
+  Cloak_10 = 'Плащ +10',
+
+  TalismanOfAden = 'Талисман Адена',
+  TalismanOfAden_1 = 'Талисман Адена +1',
+  TalismanOfAden_2 = 'Талисман Адена +2',
+  TalismanOfAden_3 = 'Талисман Адена +3',
+  TalismanOfAden_4 = 'Талисман Адена +4',
+  TalismanOfAden_5 = 'Талисман Адена +5',
+  TalismanOfAden_6 = 'Талисман Адена +6',
+  TalismanOfAden_7 = 'Талисман Адена +7',
+  TalismanOfAden_8 = 'Талисман Адена +8',
+  TalismanOfAden_9 = 'Талисман Адена +9',
+  TalismanOfAden_10 = 'Талисман Адена +10',
+
+  TalismanOfEva = 'Талисман Евы',
+  TalismanOfEva_1 = 'Талисман Евы +1',
+  TalismanOfEva_2 = 'Талисман Евы +2',
+  TalismanOfEva_3 = 'Талисман Евы +3',
+  TalismanOfEva_4 = 'Талисман Евы +4',
+  TalismanOfEva_5 = 'Талисман Евы +5',
+  TalismanOfEva_6 = 'Талисман Евы +6',
+  TalismanOfEva_7 = 'Талисман Евы +7',
+  TalismanOfEva_8 = 'Талисман Евы +8',
+  TalismanOfEva_9 = 'Талисман Евы +9',
+  TalismanOfEva_10 = 'Талисман Евы +10',
+
+  HeroCrown = 'Венец Героя',
+  HeroCrown_1 = 'Венец Героя +1',
+  HeroCrown_2 = 'Венец Героя +2',
+  HeroCrown_3 = 'Венец Героя +3',
+  HeroCrown_4 = 'Венец Героя +4',
+  HeroCrown_5 = 'Венец Героя +5',
+  HeroCrown_6 = 'Венец Героя +6',
+  HeroCrown_7 = 'Венец Героя +7',
+  HeroCrown_8 = 'Венец Героя +8',
+  HeroCrown_9 = 'Венец Героя +9',
+  HeroCrown_10 = 'Венец Героя +10',
+
+  DragonBelt = 'Пояс Дракона',
+  DragonBelt_1 = 'Пояс Дракона +1',
+  DragonBelt_2 = 'Пояс Дракона +2',
+  DragonBelt_3 = 'Пояс Дракона +3',
+  DragonBelt_4 = 'Пояс Дракона +4',
+  DragonBelt_5 = 'Пояс Дракона +5',
+  DragonBelt_6 = 'Пояс Дракона +6',
+  DragonBelt_7 = 'Пояс Дракона +7',
+  DragonBelt_8 = 'Пояс Дракона +8',
+  DragonBelt_9 = 'Пояс Дракона +9',
+  DragonBelt_10 = 'Пояс Дракона +10',
+
+  DragonEssence = 'Эссенция Дракона',
+  DragonVarnish = 'Лак Дракона',
+  DragonPendantOne = 'Подвеска Дракона Ур.1',
+  DragonPendantOne_1 = 'Подвеска Дракона Ур.1 +1',
+  DragonPendantOne_2 = 'Подвеска Дракона Ур.1 +2',
+  DragonPendantOne_3 = 'Подвеска Дракона Ур.1 +3',
+  DragonPendantOne_4 = 'Подвеска Дракона Ур.1 +4',
+  DragonPendantOne_5 = 'Подвеска Дракона Ур.1 +5',
+
+  DragonPendantTwo = 'Подвеска Дракона Ур.2',
+  DragonPendantTwo_1 = 'Подвеска Дракона Ур.2 +1',
+  DragonPendantTwo_2 = 'Подвеска Дракона Ур.2 +2',
+  DragonPendantTwo_3 = 'Подвеска Дракона Ур.2 +3',
+  DragonPendantTwo_4 = 'Подвеска Дракона Ур.2 +4',
+  DragonPendantTwo_5 = 'Подвеска Дракона Ур.2 +5',
+
+  DragonPendantThree = 'Подвеска Дракона Ур.3',
+  DragonPendantThree_1 = 'Подвеска Дракона Ур.3 +1',
+  DragonPendantThree_2 = 'Подвеска Дракона Ур.3 +2',
+  DragonPendantThree_3 = 'Подвеска Дракона Ур.3 +3',
+  DragonPendantThree_4 = 'Подвеска Дракона Ур.3 +4',
+  DragonPendantThree_5 = 'Подвеска Дракона Ур.3 +5',
+
+  DragonPendantFour = 'Подвеска Дракона Ур.4',
+  DragonPendantFour_1 = 'Подвеска Дракона Ур.4 +1',
+  DragonPendantFour_2 = 'Подвеска Дракона Ур.4 +2',
+  DragonPendantFour_3 = 'Подвеска Дракона Ур.4 +3',
+  DragonPendantFour_4 = 'Подвеска Дракона Ур.4 +4',
+  DragonPendantFour_5 = 'Подвеска Дракона Ур.4 +5',
+
+  TalismanOfAuthorityFragment = 'Осколок Талисмана Властителя',
+  TalismanOfAuthority_1 = 'Талисман Властителя Ур.1',
+  TalismanOfAuthority_2 = 'Талисман Властителя Ур.2',
+  TalismanOfAuthority_3 = 'Талисман Властителя Ур.3',
+  TalismanOfAuthority_4 = 'Талисман Властителя Ур.4',
+  TalismanOfAuthority_5 = 'Талисман Властителя Ур.5',
+
+  NobleTalismanOfAuthority_1 = 'Талисман Благородного Властителя Ур.1',
+  NobleTalismanOfAuthority_2 = 'Талисман Благородного Властителя Ур.2',
+  NobleTalismanOfAuthority_3 = 'Талисман Благородного Властителя Ур.3',
+  NobleTalismanOfAuthority_4 = 'Талисман Благородного Властителя Ур.4',
+  NobleTalismanOfAuthority_5 = 'Талисман Благородного Властителя Ур.5',
+
+  RadiantTalismanOfAuthority_1 = 'Искрящийся Талисман Властителя Ур.1',
+  RadiantTalismanOfAuthority_2 = 'Искрящийся Талисман Властителя Ур.2',
+  RadiantTalismanOfAuthority_3 = 'Искрящийся Талисман Властителя Ур.3',
+  RadiantTalismanOfAuthority_4 = 'Искрящийся Талисман Властителя Ур.4',
+  RadiantTalismanOfAuthority_5 = 'Искрящийся Талисман Властителя Ур.5',
+
+  Rune_1 = 'Руна Ур.1',
+  Rune_2 = 'Руна Ур.2',
+  Rune_3 = 'Руна Ур.3',
+  Rune_4 = 'Руна Ур.4',
+  Rune_5 = 'Руна Ур.5',
+  Rune_6 = 'Руна Ур.6',
+  Rune_7 = 'Руна Ур.7',
+  Rune_8 = 'Руна Ур.8',
+  Rune_9 = 'Руна Ур.9',
+  Rune_10 = 'Руна Ур.10',
+  Rune_11 = 'Руна Ур.11',
+  Rune_12 = 'Руна Ур.12',
+  Rune_13 = 'Руна Ур.13',
+
+  AgathionCoupon = 'Билет Обмена Агатиона',
+  AgathionEnchantScroll = 'Свиток: Модифицировать Агатиона',
+  Agathion = 'Агатион',
+  Agathion_1 = 'Агатион +1',
+  Agathion_2 = 'Агатион +2',
+  Agathion_3 = 'Агатион +3',
+  Agathion_4 = 'Агатион +4',
+  Agathion_5 = 'Агатион +5',
+  Agathion_6 = 'Агатион +6',
+  Agathion_7 = 'Агатион +7',
+  Agathion_8 = 'Агатион +8',
+  Agathion_9 = 'Агатион +9',
+  Agathion_10 = 'Агатион +10',
+
+  EvaScrollEnchantHairAccessory = 'Свиток Евы: Модифицировать Головной Убор',
+  Hat = 'Головной Убор',
+  Hat_1 = 'Головной Убор +1',
+  Hat_2 = 'Головной Убор +2',
+  Hat_3 = 'Головной Убор +3',
+  Hat_4 = 'Головной Убор +4',
+  Hat_5 = 'Головной Убор +5',
+  Hat_6 = 'Головной Убор +6',
+  Hat_7 = 'Головной Убор +7',
+  Hat_8 = 'Головной Убор +8',
+  Hat_9 = 'Головной Убор +9',
+  Hat_10 = 'Головной Убор +10',
+
+  // стихийное ожерелье
+  BlessedSpiritsVarnish = 'Благословенный Лак Духов',
+  IgnisNecklace = 'Ожерелье Игниса',
+  IgnisNecklace_1 = 'Ожерелье Игниса +1',
+  IgnisNecklace_2 = 'Ожерелье Игниса +2',
+  IgnisNecklace_3 = 'Ожерелье Игниса +3',
+  IgnisNecklace_4 = 'Ожерелье Игниса +4',
+  IgnisNecklace_5 = 'Ожерелье Игниса +5',
+  IgnisNecklace_6 = 'Ожерелье Игниса +6',
+  IgnisNecklace_7 = 'Ожерелье Игниса +7',
+  IgnisNecklace_8 = 'Ожерелье Игниса +8',
+  IgnisNecklace_9 = 'Ожерелье Игниса +9',
+  IgnisNecklace_10 = 'Ожерелье Игниса +10',
+
+  EWB = 'Свиток: Модифицировать Оружие (Ранг B)',
+  BOP = 'Лук Угрозы',
+  BOP_1 = 'Лук Угрозы +1',
+  BOP_2 = 'Лук Угрозы +2',
+  BOP_3 = 'Лук Угрозы +3',
+  BOP_4 = 'Лук Угрозы +4',
+  BOP_5 = 'Лук Угрозы +5',
+  BOP_6 = 'Лук Угрозы +6',
+  BOP_7 = 'Лук Угрозы +7',
+  BOP_8 = 'Лук Угрозы +8',
+  BOP_9 = 'Лук Угрозы +9',
+  BOP_10 = 'Лук Угрозы +10',
+  BOP_11 = 'Лук Угрозы +11',
+  BOP_12 = 'Лук Угрозы +12',
+  BOP_13 = 'Лук Угрозы +13',
+  BOP_14 = 'Лук Угрозы +14',
+  BOP_15 = 'Лук Угрозы +15',
+  BOP_16 = 'Лук Угрозы +16',
+}
